requirement.txt
pandas>=1.3.0
numpy>=1.21.0
pyyaml>=6.0
openpyxl>=3.0.0
plotly>=5.3.0
pytest>=6.2.0
python-dotenv>=0.19.0